# proviemplea/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea",
        "description": "Documentación oficial de la API ProviEmplea.\n\n**Rate Limiting:** Máximo 60 solicitudes por minuto por IP.\n**Caché:** Los endpoints GET tienen caché de 5 minutos.\n**Tiempos de respuesta esperados:** GET ~100ms | POST ~200ms | PUT ~150ms | DELETE ~100ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/contactos": {
            "get": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Listar solicitudes de contacto",
                "description": "Retorna todas las solicitudes de contacto con datos de empresa y persona relacionadas.",
                "operationId": "f78385f042cdd9221c11a42c20c170b5",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado: pendiente, aceptado, rechazado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "pendiente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contacto"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "772a1622-a40d-53e6-c928-668877662222",
                                        "empresa_id": "661f9511-f39c-42d5-b817-557766551111",
                                        "persona_id": "550e8400-e29b-41d4-a716-446655440000",
                                        "estado": "pendiente",
                                        "notas_admin": null
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Crear solicitud de contacto",
                "description": "Una empresa solicita contactar a un talento. El estado inicial es pendiente.",
                "operationId": "a5f0e367dc7052ea351cc12ed7776b28",
                "requestBody": {
                    "description": "Datos de la solicitud de contacto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contacto"
                            },
                            "example": {
                                "empresa_id": "661f9511-f39c-42d5-b817-557766551111",
                                "persona_id": "550e8400-e29b-41d4-a716-446655440000",
                                "estado": "pendiente",
                                "notas_admin": "Contacto iniciado desde búsqueda por rubro Tecnología"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contacto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación — empresa_id o persona_id no existen"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/contactos/{id}": {
            "get": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Obtener solicitud de contacto por ID",
                "description": "Retorna los datos de una solicitud junto con la empresa y talento relacionados.",
                "operationId": "79a453f2c4567871d8ae3e94090d169b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del contacto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "772a1622-a40d-53e6-c928-668877662222"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contacto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Actualizar solicitud de contacto (PUT)",
                "description": "Permite actualizar el estado o las notas de una solicitud existente.",
                "operationId": "c504bb803bc05172b84d0aa71bfa28fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del contacto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "772a1622-a40d-53e6-c928-668877662222"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contacto"
                            },
                            "example": {
                                "estado": "aceptado",
                                "notas_admin": "Empresa verificada y talento notificado"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contacto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contacto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Eliminar solicitud de contacto",
                "description": "Elimina definitivamente una solicitud de contacto de la base de datos.",
                "operationId": "1b6ab274c05e69bea20c30ac970d5ab3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del contacto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "772a1622-a40d-53e6-c928-668877662222"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Contacto": {
                "required": [
                    "empresa_id",
                    "persona_id",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "772a1622-a40d-53e6-c928-668877662222"
                    },
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "661f9511-f39c-42d5-b817-557766551111"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "estado": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "notas_admin": {
                        "type": "string",
                        "example": "Empresa verificada, contacto válido"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}